+++ ARNavigationControllerDelegateProxy.m
@@ -1,3 +1,4 @@
+#import <UIKit/UIKit.h>
 #import "ARNavigationControllerDelegateProxy.h"
 
 @interface ARNavigationControllerDelegateProxy ()
@@ -8,7 +9,7 @@
 
 @implementation ARNavigationControllerDelegateProxy
 
-- (instancetype)initWithAnalyticsDelegate:(id<UINavigationControllerDelegate>)analyticsDelegate {
+- (instancetype)initWithAnalyticsDelegate:(NSObject<UINavigationControllerDelegate> *)analyticsDelegate {
     // No need for [super init], it does not exist
     _analyticsDelegate = analyticsDelegate;
     return self;