+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added super properties that get sent with all event methods
 * Added support for AppsFlyer ( thanks @cdzombak )
 * Added support for Branch.io ( thanks @scotthasbrouck )
+* Updated Adjust provider to support Adjust 4.0 ( thanks @HeEAaD )
 
 ## Version 2.9.0