+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added support for AppsFlyer ( thanks @cdzombak )
 * Added support for Branch.io ( thanks @scotthasbrouck )
 * Updated Adjust provider to support Adjust 4.0 ( thanks @HeEAaD )
+* Fixes warnings generated from ARNavigationControllerDelegateProxy (thanks @ashfurrow )
 
 ## Version 2.9.0